requirements.txt
flask==3.0.3
flask-cors==4.0.1
pymongo[srv]==4.8.0
bcrypt==4.2.0
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
google-generativeai==0.7.2
numpy==1.26.4

Flask==2.3.3
Flask-Session==0.5.0
pymongo==4.5.0
google-generativeai==0.3.2
bcrypt==4.0.1
python-dotenv==1.0.0
requests==2.31.0